+++ applications/convection_diffusion_application/tests/deprecated_thermal_coupling_test.py
